SK-601: fix credential to DID matching
