+++ src/views/SignQuote/SignQuote.test.tsx
@@ -1,11 +1,18 @@
 import { render } from '../../testing/testing';
 import { mockTerms } from '../../utilities/cTypes/cTypes.mock';
 import { PopupTestProvider } from '../../utilities/popups/PopupTestProvider';
+import { parseDidUrl } from '../../utilities/did/did';
 import { identitiesMock as identities } from '../../utilities/identities/IdentitiesProvider.mock';
 import { waitForGetPassword } from '../../channels/SavedPasswordsChannels/SavedPasswordsChannels.mock';
 import { paths } from '../paths';
 
 import { SignQuote } from './SignQuote';
+
+jest.mock('../../utilities/did/did');
+
+(parseDidUrl as jest.Mock).mockReturnValue({
+  fullDid: 'did:kilt:4rrkiRTZgsgxjJDFkLsivqqKTqdUTuxKk3FX3mKFAeMxsR51',
+});
 
 describe('SignQuote', () => {
   it('should render', async () => {
